Added TaskQueues to process the request to delete all the books
