+++ src/main/java/com/fullcreative/bms/controllers/LoginController.java
@@ -1,4 +1,4 @@
-package com.fullcreative.controllers;
+package com.fullcreative.bms.controllers;
 
 import java.io.IOException;
 import java.util.HashMap;
@@ -11,8 +11,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import com.fullcreative.services.AuthServices;
-import com.fullcreative.utilities.UserControllerUtilities;
+import com.fullcreative.bms.services.AuthServices;
+import com.fullcreative.bms.utilities.UserControllerUtilities;
 import com.google.gson.Gson;
 
 
@@ -36,7 +36,7 @@ public class LoginController extends HttpServlet {
 				HttpSession newSession = request.getSession(true);
 				newSession.setAttribute("isLoggedIn", "yes");
 				System.out.println(newSession.getAttribute("isLoggedIn"));
-				newSession.setMaxInactiveInterval(5 * 60);
+				newSession.setMaxInactiveInterval(60 * 60);
 			}
 			response.setContentType("application/json");
 			response.getWriter().print(respopnseAsJson);
