+++ src/main/webapp/WEB-INF/web.xml
@@ -19,13 +19,16 @@
  
 <filter>
  	<filter-name>AuthenticationFilter</filter-name>
- 	<filter-class>com.fullcreative.filters.AuthFilter</filter-class>
+ 	<filter-class>com.fullcreative.bms.filters.AuthFilter</filter-class>
  </filter>
  
  <filter-mapping>
  	<filter-name>AuthenticationFilter</filter-name>
  	<url-pattern>/books</url-pattern>
  	<url-pattern>/books/*</url-pattern>
+ 	<url-pattern>/images</url-pattern>
+ 	<url-pattern>/images/*</url-pattern>
+ 	<url-pattern>/users</url-pattern>
  </filter-mapping>
