+++ src/main/webapp/index.html
@@ -7,6 +7,8 @@
     <title>Book Manager</title>
     <link rel="icon" href="./images/book-manager-logo.svg" type="image/svg">
     <link rel="stylesheet" href="./styles/auth.css">
+    <script src="./js/objects.js"></script>
+    <script src="./js/utilityFunctions.js"></script>
     <script src="./js/login.js" defer></script>
 </head>
