+++ src/main/webapp/signup.html
@@ -7,6 +7,7 @@
     <title>Book Manager</title>
     <link rel="icon" href="./images/book-manager-logo.svg" type="image/svg">
     <link rel="stylesheet" href="./styles/auth.css">
+    <script src="./js/objects.js"></script>
     <script src="./js/signup.js" defer></script>
 </head>
